Modified item sync on scene load
Commented this in the first place because I would instead sync object as they instantiate during scene load, but apparently the scene load instantiation doesn't go through the same instantiation pipeline as normal so I still have to do it when we arrive in the scene. The sync during load stays because objects may still be instantiated during load through the normal pipeline.