Made tracked item pos/rot completely dependent on parentage
Offset attachments was probably being caused by non controller receiving a local position as data position but we had yet to receive the parent or have the parent instance. When we then receive the parent, we parented the item but we didn't receive more updates during which we would update the position and correctly set it to local, so the items were never positioned properly.

TrackedITems will now set their pos/rot accordingly when they attach to a parent