Fixed attachment scaling issue
By ensuring scale of object we parent to is consistent across the network
Note that this should always be the case, if it isn't, like the case for which we introduced this fix, it was incorrect to begin with. This just makes sure that if there are such errors in assets, they will be corrected by H3MP as needed.