Fixed tracked objects sometimes not being instantiated
Specifically in the case of tracked objects that have changed scene/instance. Consider the case where server hosts in a scene, client joins in another scene, server moves to the other scene. The host's body is switched scene, but the client never had the data for that body in the first place since they were never in the same scene as it before now. The server never sends the full data of the body to the client, only its new scene, but the client, missing the body data, ends up ignoring it. We now send full data instead of just the scene, so that the player may add the data to their array if necessary.